ui added singleworkout base page, server added checkin entity, viewable for workouts/exercises

+++ server/stevens-training/src/controllers/exercise.controller.ts
@@ -17,8 +17,15 @@ export class ExerciseController {
 
   @Get(':id')
   async getExercise(@Req() req, @Param('id') exerciseId: string): Promise<Exercise> {
-      this.logger.log(`Get exercise ${exerciseId} by user ${req.user.userName}`);
-      return this.exerciseService.findOne(exerciseId);
+    const exercise = await this.exerciseService.findOne(exerciseId);
+    if(exercise.user.id !== req.user.id) {
+      if(!exercise.viewable){
+        this.logger.log(`Exercise is not viewable by other users, requested by ${req.user.userName}, exercise is owned by user ID: ${exercise.user.userName}`);
+        return null;
+      }
+    }
+    this.logger.log(`Get exercise ${exerciseId} by user ${req.user.userName}`);
+    return this.exerciseService.findOne(exerciseId);
   }
 
   @Post('create')
@@ -29,6 +36,11 @@ export class ExerciseController {
 
   @Delete(':id')
   async removeExercise(@Req() req, @Param('id') exerciseId: string): Promise<void> {
+    const exercise = await this.exerciseService.findOne(exerciseId);
+    if(exercise.user.id !== req.user.id) {
+      this.logger.log(`Exercise is not editable by other users, requested by ${req.user.userName}, exercise is owned by user ID: ${exercise.user.userName}`);
+      return null;
+    }
     this.logger.log(`Remove exercise called by user ${req.user.userName}`);
     return this.exerciseService.delete(exerciseId);
   }
